testing admin page

+++ views/admin.view.php
@@ -1,6 +1,3 @@
-
-
-
 <body>  
 <br><br>    
 <h3>Add New Project</h3>
@@ -19,6 +16,14 @@
         <textarea id="content" name="content" rows="4" style ="width: 100%" maxlength="600"></textarea>
     </form>
     <h3>Existing Projects</h3>
+    <form method="POST">
+        <label for="filter">Filter:</label>
+        <select id="filter" name="filter">
+            <option value="all">All Projects</option>
+            <option value="projects">Projects Only</option>
+        </select>
+        <button type="submit">Apply Filter</button>
+    </form>
     <table border="1">
         <thead>
             <tr>
@@ -33,17 +38,20 @@
         </thead>
         <tbody>
         <?php
+            $filter = $_POST['filter'] ?? 'all';
             foreach ($data as $projects) 
             {
-                echo "<tr>";
-                echo "<td>{$projects['id']}</td>";
-                echo "<td>{$projects['page']}</td>";
-                echo "<td>{$projects['title']}</td>";
-                echo "<td>{$projects['content']}</td>";
-                echo "<td>{$projects['languages']}</td>";
-                echo "<td>{$projects['link']}</td>";
-                echo "<td><a href='" .  URL_ROOT . "/scripts/edit.php?id={$projects['id']}'>Edit</a> | <a href='" .  URL_ROOT . "/scripts/delete.php?id={$projects['id']}' onclick='return confirm(\"Are you sure you want to delete this project?\");'>Delete</a></td>";
-                echo "</tr>";
+                if ($filter == 'all' || ($filter == 'projects' && $projects['title'] == 'projects')) {
+                    echo "<tr>";
+                    echo "<td>{$projects['id']}</td>";
+                    echo "<td>{$projects['page']}</td>";
+                    echo "<td>{$projects['title']}</td>";
+                    echo "<td>{$projects['content']}</td>";
+                    echo "<td>{$projects['languages']}</td>";
+                    echo "<td>{$projects['link']}</td>";
+                    echo "<td><a href='" .  URL_ROOT . "/scripts/edit.php?id={$projects['id']}'>Edit</a> | <a href='" .  URL_ROOT . "/scripts/delete.php?id={$projects['id']}' onclick='return confirm(\"Are you sure you want to delete this project?\");'>Delete</a></td>";
+                    echo "</tr>";
+                }
             }
             include_once(APP_ROOT . '/views/nav.view.php')
         ?>
